rename from basic workflow to chat with tools
