--- template_langgraph/tasks/draw_basic_workflow_agent_mermaid_png.py
@@ -1,11 +1,0 @@
-import sys
-
-from template_langgraph.agents.basic_workflow_agent.agent import BasicWorkflowAgent
-
-if __name__ == "__main__":
-    png_path = "data/basic_workflow_agent.png"
-    if len(sys.argv) > 1:
-        png_path = sys.argv[1]
-
-    with open(png_path, "wb") as f:
-        f.write(BasicWorkflowAgent().draw_mermaid_png())
